+++ rod/src/main/java/com/example/rod/question/service/QuestionServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.rod.question.service;
 
-import com.example.rod.answer.dto.AnswerResponseDto;
+import com.example.rod.answer.dto.AnswerWithCommentsDto;
 import com.example.rod.answer.entity.Answer;
 import com.example.rod.answer.repository.AnswerRepository;
 import com.example.rod.comment.dto.CommentResponseDto;
@@ -9,24 +9,17 @@ import com.example.rod.comment.repository.CommentRepository;
 import com.example.rod.question.dto.*;
 import com.example.rod.question.entity.Question;
 import com.example.rod.question.repository.QuestionRepository;
-import com.example.rod.security.details.UserDetailServiceImpl;
 import com.example.rod.security.details.UserDetailsImpl;
-import com.example.rod.security.jwt.JwtUtil;
 import com.example.rod.user.entity.User;
 import com.example.rod.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.autoconfigure.neo4j.Neo4jProperties;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
@@ -48,22 +41,10 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     @Transactional
     public void createQuestion(QuestionRequest questionRequest, UserDetailsImpl userDetails){
-//        String token = request.getParameter("Authentication");
-//        User user = userDetails.getUser();
         User user = userDetails.getUser();
-//        User user = userRepository.findById(userDetails).orElseThrow(
-//                () -> new UsernameNotFoundException("사용자를 찾을 수 없습니다.")
-//        );
-
-//        UserDetails newUserDetails= userDe
-//        tailsService.loadUserByUsername(username);
-//        User user = userRepository.findById(userId).orElseThrow(
-//                () -> new IllegalArgumentException("찾는 유저가 없습니다.")
-//        );
         Question question = Question.builder()
                 .title(questionRequest.getTitle())
                 .content(questionRequest.getContent())
-//                .user(newUserDetails.u)
                 .user(user)
                 .build();
         questionRepository.save(question);
@@ -71,19 +52,17 @@ public class QuestionServiceImpl implements QuestionService {
 
     @Override
     @Transactional
-    public GetQuestionsResponse getMyQuestions(Long userId, Pageable pageable, int page) {
+    public GetQuestionsResponse getMyQuestions(UserDetailsImpl userDetails,Pageable pageable, int page) {
 
-        User user = userRepository.findById(userId).orElseThrow
-                (()-> new IllegalArgumentException("해당 아이디의 유저가 없습니다."));
+        User user = userDetails.getUser();
 
         Page<Question> questionList = questionRepository.findAllByUser(user, pageable.withPage(page-1));
 
         List<QuestionResponse> questionResponseList = new ArrayList<>();
 
         for (Question question : questionList) {
-            questionResponseList.add(new QuestionResponse(question));
+            questionResponseList.add(new QuestionResponse(question.getQuestionId(), question.getTitle(), question.getContent()));
         }
-
         return new GetQuestionsResponse(page, questionResponseList);
     }
 
@@ -98,7 +77,7 @@ public class QuestionServiceImpl implements QuestionService {
 
 
         for (Question question : questionList) {
-            questionResponseList.add(new QuestionResponse(question));
+            questionResponseList.add(new QuestionResponse(question.getQuestionId(), question.getTitle(), question.getContent()));
         }
 
         return new GetQuestionsResponse(page, questionResponseList);
@@ -111,7 +90,7 @@ public class QuestionServiceImpl implements QuestionService {
         Question question = questionRepository.findById(questionId).orElseThrow
                 (() -> new IllegalArgumentException("해당 아이디의 질문이 없습니다."));
 
-        List<AnswerResponseDto> answerWithComments = new ArrayList<>();
+        List<AnswerWithCommentsDto> answerWithComments = new ArrayList<>();
 
 
         // 1. AnswerList 페이징 처리해서 뽑아온다.
@@ -127,8 +106,8 @@ public class QuestionServiceImpl implements QuestionService {
                 CommentResponseDto commentResponseDto = new CommentResponseDto(comment.getId(), comment.getContent());
                 comments.add(commentResponseDto);
             }
-            AnswerResponseDto answerResponseDto = new AnswerResponseDto(answer.getId(), answer.getContent(), answer.getLikes(), comments);
-            answerWithComments.add(answerResponseDto);
+            AnswerWithCommentsDto answerWithCommentsDto = new AnswerWithCommentsDto(answer.getId(), answer.getContent(), answer.getLikes(), comments);
+            answerWithComments.add(answerWithCommentsDto);
         }
 
         QuestionWithAnswersResponse questionWithAnswersResponse = new QuestionWithAnswersResponse(question.getTitle(), question.getContent(), answerWithComments);
@@ -138,26 +117,53 @@ public class QuestionServiceImpl implements QuestionService {
 
     @Override
     @Transactional
-    public void changeQuestionTitle(Long questionId, PatchQuestionTitleRequest patchQuestionTitleRequest) {
+    public void changeQuestionTitle(Long questionId, PatchQuestionTitleRequest patchQuestionTitleRequest, UserDetailsImpl userDetails) {
         Question question = questionRepository.findById(questionId).orElseThrow
                 (()-> new IllegalArgumentException("해당 아이디의 질문이 없습니다."));
-        question.editTitle(patchQuestionTitleRequest.getTitle());
+
+        User user = userDetails.getUser();
+
+        if(question.isOwnedBy(user)){
+            question.editTitle(patchQuestionTitleRequest.getTitle());
+        } else {
+            throw new IllegalArgumentException("수정 권한이 없는 유저입니다.");
+        }
     }
 
 
     @Override
     @Transactional
-    public void changeQuestionContent(Long questionId, PatchQuestionContentRequest patchQuestionContentRequest){
+    public void changeQuestionContent(Long questionId, PatchQuestionContentRequest patchQuestionContentRequest, UserDetailsImpl userDetails){
         Question question = questionRepository.findById(questionId).orElseThrow
                 (()-> new IllegalArgumentException("해당 아이디의 질문이 없습니다."));
-        question.editContent(patchQuestionContentRequest.getContent());
+
+        User user = userDetails.getUser();
+
+        if(question.isOwnedBy(user)){
+            question.editContent(patchQuestionContentRequest.getContent());
+        } else {
+            throw new IllegalArgumentException("수정 권한이 없는 유저입니다.");
+        }
+
+
     }
 
 
     @Override
     @Transactional
-    public void deleteQuestion(Long questionId) {
-        questionRepository.deleteById(questionId);
+    public void deleteQuestion(Long questionId, UserDetailsImpl userDetails) {
+
+        Question question = questionRepository.findById(questionId).orElseThrow(
+                () -> new IllegalArgumentException("해당 아이디의 질문이 없습니다.")
+        );
+
+        User user = userDetails.getUser();
+
+        if(question.isOwnedBy(user)){
+            questionRepository.deleteById(questionId);
+        } else {
+            throw new IllegalArgumentException("삭제 권한이 없는 유저입니다.");
+        }
     }
